Move Grade.java file.
